backend/requirements.txt
Django>=5.0,<6
djangorestframework>=3.15
django-cors-headers>=4.4
psycopg[binary]>=3.1
python-dotenv>=1.0
Pillow>=10.0
